+++ __tests__/at-es5-test.js
@@ -2,7 +2,7 @@
 
 const selfTimer = require("../src/build/es5/intermediate/selftimer");
 
-describe("SelfTimer.js *test group Test", () => {
+describe("SelfTimer.js *at group Test", () => {
   const _str = "working";
 
   it("SelfTimer at().Between Test", () => {
@@ -61,9 +61,7 @@ describe("SelfTimer.js *test group Test", () => {
 
     const task = new selfTimer(date);
 
-    const callback = task.at().HoursBetween(0, 23, function() {
-      return _str;
-    });
+    const callback = task.at().HoursBetween(0, 23, () => _str);
 
     expect(callback).toBe(_str);
     expect(task.at().HoursBetween(0, 23)).toBeTruthy();
@@ -76,9 +74,37 @@ describe("SelfTimer.js *test group Test", () => {
 
     const task = new selfTimer(date);
 
-    const callback = task.at().HourSelects(hours, () => {
-      return _str;
-    });
+    const callback = task.at().HourSelects(hours, () => _str);
+
+    expect(callback).toBe(_str);
+  });
+
+  it("at().Min Test", () => {
+    const date = new Date("Tue Oct 17 2017 17:14:12 GMT+0900 (JST)");
+
+    const task = new selfTimer(date);
+
+    const callback = task.at().Min(14, () => _str);
+
+    expect(callback).toBe(_str);
+  });
+
+  it("at().MinBetween Test", () => {
+    let date = new Date("Tue Oct 17 2017 17:14:12 GMT+0900 (JST)");
+
+    const task = new selfTimer(date);
+
+    const callback = task.at().MinBetween(14, 17, () => _str);
+
+    expect(callback).toBe(_str);
+  });
+
+  it("at().MinSelects Test", () => {
+    let date = new Date("Tue Oct 17 2017 17:14:12 GMT+0900 (JST)");
+
+    const task = new selfTimer(date);
+
+    const callback = task.at().MinSelects([14, 17], () => _str);
 
     expect(callback).toBe(_str);
   });