+++ src/methods/es5/at.js
@@ -9,6 +9,7 @@ SelfTimer.prototype.at = function() {
   var _Current = this.D;
   var _day = this.D.getDay();
   var _hour = this.D.getHours();
+  var _minute = this.D.getMinutes();
   var _msg = this.messages();
   var _h = this.helpers();
   var _fmt = this.formats();
@@ -72,6 +73,74 @@ SelfTimer.prototype.at = function() {
       } // ! Exception
     } // ! if()
   }; // ! Unless()
+
+  var Min = function(minute, task) {
+    if (_h.__checkIsValid(task)) {
+      var time = parseInt(minute);
+
+      try {
+        if (time > 59) throw _msg.startMin;
+
+        if (time === _minute) return task !== undefined ? task() : true;
+      } catch (e) {
+        console.error(e);
+        return;
+      } // ! Exception
+    } // ! if()
+
+  } // ! Minute()
+
+  var MinBetween = function(from, to, task) {
+    if (_h.__checkIsValid(task)) {
+      var start = parseInt(from);
+      var end = parseInt(to);
+
+      try {
+        if (start > 59) throw _msg.startMin;
+
+        if (end > 59) throw _msg.endMin;
+
+        var arr = _h.__range(end, end - start);
+        arr.push(from);
+
+        if (_h.__contains(arr, _minute)) {
+          return task !== undefined ? task() : true;
+        } // ! if()
+      } catch (e) {
+        console.error(e);
+        return;
+      } // ! Exception
+    } // ! if()
+  }; // ! MinutesBetween()
+
+
+  var MinSelects = function(minutes, task) {
+    if (_h.__checkIsValid(task)) {
+      try {
+        if (!Array.isArray(minutes)) throw _msg.isNotArray;
+
+        // check array elements if numberic
+        if (!minutes.some(isNaN) != true) throw _msg.minFormat;
+
+        var array = minutes.map(function(res) {
+          // convert elemetnts to Integer in array
+          return parseInt(res);
+        });
+
+        array.map(function(res) {
+          // check elements if valid minute format
+          if (res > 59) throw _msg.isNotValidMin;
+        });
+
+        if (_h.__contains(array, _minute)) {
+          return task !== undefined ? task() : true;
+        } // ! if
+      } catch (e) {
+        console.error(e);
+        return;
+      } // ! Exception
+    } // ! if()
+  }; // ! MinuteSelects()
 
   /**
    * [ at().Hour description]
@@ -161,6 +230,9 @@ SelfTimer.prototype.at = function() {
   var REGISTER = {
     Between: Between,
     Unless: Unless,
+    Min: Min,
+    MinBetween: MinBetween,
+    MinSelects: MinSelects,
     Hour: Hour,
     HoursBetween: HoursBetween,
     HourSelects: HourSelects